Add eventhandlers to the 'add book' dialog form
Adds eventhandlers to the dialog form used to add books to the library
display. Also adds additional buttons to the form used to cancel the
form and to submit a new book entry. Finally adds a new resetLibrary
function to clear the library DOM object's children in order to prevent
duplicate nodes from being displayed.